move all stats functions to gen_helper.py and delete record_stats.py
